authorization checks on back-end

+++ src/actions/actions.ts
@@ -1,18 +1,35 @@
 'use server';
 
 import { revalidatePath } from 'next/cache';
+import bcrypt from 'bcryptjs';
 
 import prisma from '@/lib/db';
 import { sleep } from '@/lib/utils';
 import { petFormSchema, petIdSchema } from '@/lib/validations';
-import { signIn, signOut } from '@/lib/auth';
+import { auth, signIn, signOut } from '@/lib/auth';
+import { redirect } from 'next/navigation';
+import { checkAuth } from '@/lib/server-utils';
 
 // --- user actions ---
 
-export async function logIn(formData: FormData) {
-  const authData = Object.fromEntries(formData.entries());
+export async function signUp(formData: FormData) {
+  const hashedPassword = await bcrypt.hash(
+    formData.get('password') as string,
+    10,
+  );
 
-  await signIn('credentials', authData);
+  await prisma.user.create({
+    data: {
+      email: formData.get('email') as string,
+      hashedPassword,
+    },
+  });
+
+  await signIn('credentials', formData);
+}
+
+export async function logIn(formData: FormData) {
+  await signIn('credentials', formData);
 }
 
 export async function logOut() {
@@ -23,6 +40,7 @@ export async function logOut() {
 
 export async function addPet(pet: unknown) {
   await sleep(1000);
+  const session = await checkAuth();
 
   const validatedPet = petFormSchema.safeParse(pet);
   if (!validatedPet.success) {
@@ -33,7 +51,14 @@ export async function addPet(pet: unknown) {
 
   try {
     await prisma.pet.create({
-      data: validatedPet.data,
+      data: {
+        ...validatedPet.data,
+        user: {
+          connect: {
+            id: session.user?.id,
+          },
+        },
+      },
     });
   } catch (error) {
     return {
@@ -46,7 +71,10 @@ export async function addPet(pet: unknown) {
 
 export async function editPet(petId: unknown, newPetData: unknown) {
   await sleep(1000);
+  // authentication check
+  const session = await checkAuth();
 
+  // validation check
   const validatedPetId = petIdSchema.safeParse(petId);
   const validatedPet = petFormSchema.safeParse(newPetData);
 
@@ -56,6 +84,21 @@ export async function editPet(petId: unknown, newPetData: unknown) {
     };
   }
 
+  // authorization check (user owns pet)
+  const pet = await prisma.pet.findUnique({ where: { id: validatedPetId.data } });
+
+  if (!pet) {
+    return {
+      message: "pet not found",
+    }
+  }
+  if (pet.userId !== session.user.id) {
+    return {
+      message: "Not authorized",
+    }
+  }
+
+  // database mutation
   try {
     await prisma.pet.update({
       where: {
@@ -74,7 +117,10 @@ export async function editPet(petId: unknown, newPetData: unknown) {
 
 export async function checkoutPet(petId: unknown) {
   await sleep(1000);
+  // authentication check
+  const session = await checkAuth();
 
+  // validation check
   const validatedPetId = petIdSchema.safeParse(petId);
 
   if (!validatedPetId.success) {
@@ -83,6 +129,21 @@ export async function checkoutPet(petId: unknown) {
     };
   }
 
+  // authorization check (user owns pet)
+  const pet = await prisma.pet.findUnique({ where: { id: validatedPetId.data } });
+
+  if (!pet) {
+    return {
+      message: "pet not found",
+    }
+  }
+  if (pet.userId !== session.user.id) {
+    return {
+      message: "Not authorized",
+    }
+  }
+
+  // database mutation
   try {
     await prisma.pet.delete({
       where: {
